// src/resources/resultTypes/sectionList.ts
export interface Run {
	text: string;
}